+++ src/main/java/lermitage/intellij/extra/icons/IconType.java
@@ -2,6 +2,7 @@
 
 package lermitage.intellij.extra.icons;
 
+/** The type of icon: the path of a bundled icon, or a Base64 encoded user icon stored in config. */
 public enum IconType {
     /** The relative path of a bundled icon. */
     PATH,