+++ src/main/java/lermitage/intellij/extra/icons/ModelCondition.java
@@ -2,7 +2,6 @@
 
 package lermitage.intellij.extra.icons;
 
-import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.util.xmlb.annotations.OptionTag;
@@ -11,8 +10,8 @@ import lermitage.intellij.extra.icons.enablers.IconEnabler;
 import lermitage.intellij.extra.icons.enablers.IconEnablerProvider;
 import lermitage.intellij.extra.icons.enablers.IconEnablerType;
 import org.intellij.lang.annotations.Language;
+import org.jetbrains.annotations.Nullable;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -23,11 +22,8 @@ import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-@SuppressWarnings("OptionalUsedAsFieldOrParameterType")
 @Tag
 public class ModelCondition {
-
-    private static final Logger LOGGER = Logger.getInstance(ModelCondition.class);
 
     @OptionTag
     private boolean start = false;
@@ -48,8 +44,6 @@ public class ModelCondition {
     @OptionTag
     private boolean checkFacets = false;
     @OptionTag
-    private boolean checkFilesExist = false;
-    @OptionTag
     private boolean hasIconEnabler = false;
 
     @OptionTag
@@ -62,7 +56,6 @@ public class ModelCondition {
     private String regex;
     @OptionTag
     private String[] facets = new String[0];
-    private String[] filesExist = new String[0];
     private Pattern pattern;
     private IconEnablerType iconEnablerType;
 
@@ -106,25 +99,20 @@ public class ModelCondition {
         this.facets = facets;
     }
 
-    public void setFilesExist(String[] filesExist) {
-        this.checkFilesExist = true;
-        this.filesExist = filesExist;
-    }
-
     public void setIconEnablerType(IconEnablerType iconEnablerType) {
         this.hasIconEnabler = true;
         this.iconEnablerType = iconEnablerType;
     }
 
-    public boolean check(String parentName, String fileName, Optional<String> fullPath, Set<String> prjFacets, Project project) {
+    public boolean check(String parentName, String fileName, @Nullable String fullPath, Set<String> prjFacets, Project project) {
         if (!enabled) {
             return false;
         }
 
-        if (hasIconEnabler && fullPath.isPresent()) {
+        if (hasIconEnabler && fullPath != null) {
             Optional<IconEnabler> iconEnabler = IconEnablerProvider.getIconEnabler(project, iconEnablerType);
             if (iconEnabler.isPresent()) {
-                boolean iconEnabledVerified = iconEnabler.get().verify(project, fullPath.get());
+                boolean iconEnabledVerified = iconEnabler.get().verify(project, fullPath);
                 if (!iconEnabledVerified) {
                     return false;
                 } else if (iconEnabler.get().terminatesConditionEvaluation()) {
@@ -147,32 +135,6 @@ public class ModelCondition {
             }
         }
 
-        // filesExist is a pre-condition, should always be associated with other conditions
-        if (checkFilesExist && filesExist != null) {
-            String projectBasePath = project.getBasePath();
-            if (projectBasePath != null) {
-                boolean atLeastOneFileChecked = false;
-                projectBasePath = projectBasePath.replaceAll("\\\\", "/");
-                if (!projectBasePath.endsWith("/")) {
-                    projectBasePath += "/";
-                }
-                for (String fileExistName : filesExist) {
-                    try {
-                        File fileExist = new File(projectBasePath + fileExistName);
-                        if (fileExist.exists() && fileExist.isFile()) {
-                            atLeastOneFileChecked = true;
-                            break;
-                        }
-                    } catch (Exception e) {
-                        LOGGER.warn("Can't check file exists: '" + projectBasePath + fileExistName + "'", e);
-                    }
-                }
-                if (!atLeastOneFileChecked) {
-                    return false;
-                }
-            }
-        }
-
         if (checkParent) {
             if (!(start || eq || end || mayEnd)) {
                 if (parentNames.contains(parentName)) {
@@ -185,11 +147,11 @@ public class ModelCondition {
             }
         }
 
-        if (hasRegex && fullPath.isPresent()) {
+        if (hasRegex && fullPath != null) {
             if (pattern == null) {
                 pattern = Pattern.compile(regex);
             }
-            if (pattern.matcher(fullPath.get()).matches()) {
+            if (pattern.matcher(fullPath).matches()) {
                 return true;
             }
         }